[Chaos][OADP] Add case CNV-12011: Reboot the worker node during OADP Backup
